refactor(analyser): split _diagnostics.py into 7 focused sub-modules
_diagnostics.py was 1395 lines. Split by diagnostic code group:

  _diag_commands.py       W123 — unresolved commands, interp resolution
  _diag_var_lifecycle.py  W210/211/213/214/220/H300 — variable lifecycle
  _diag_var_command.py    W307/308 — $var-as-command / method validation
  _diag_branches.py       I230/231 — constant branch / unreachable arms
  _diag_channel.py        W126 — non-channel value in channel position
  _diag_ip.py             W124 — invalid IP address literals
  _diag_racy.py           IRULE4005 — racy static:: cross-event writes

_diagnostics.py is now a ~200-line orchestrator that wires
_emit_cfg_ssa_diagnostics and _emit_cfg_ssa_diagnostics_for_function
via self.* calls into the assembled Analyser mixin class.

All 7 new mixins are added to Analyser in __init__.py.
ty: 0 errors, ruff: clean, 252 tests pass.
