+++ core/analysis/_analyser/__init__.py
@@ -2,6 +2,13 @@ from __future__ import annotations
 
 from ._commands import _AnalyserCommandsMixin
 from ._core import _AnalyserBase
+from ._diag_branches import _AnalyserDiagBranchesMixin
+from ._diag_channel import _AnalyserDiagChannelMixin
+from ._diag_commands import _AnalyserDiagCommandsMixin
+from ._diag_ip import _AnalyserDiagIPMixin
+from ._diag_racy import _AnalyserDiagRacyMixin
+from ._diag_var_command import _AnalyserDiagVarCommandMixin
+from ._diag_var_lifecycle import _AnalyserDiagVarLifecycleMixin
 from ._diagnostics import _AnalyserDiagsMixin
 from ._handlers import _AnalyserHandlersMixin
 from ._oo import _AnalyserOOMixin
@@ -20,7 +27,14 @@ class Analyser(
     _AnalyserProcMixin,
     _AnalyserOOMixin,
     _AnalyserDiagsMixin,
-    _AnalyserBase,
+    _AnalyserDiagCommandsMixin,
+    _AnalyserDiagVarCommandMixin,
+    _AnalyserDiagBranchesMixin,
+    _AnalyserDiagChannelMixin,
+    _AnalyserDiagIPMixin,
+    _AnalyserDiagRacyMixin,
+    _AnalyserDiagVarLifecycleMixin,
+    _AnalyserBase,  # last — must follow all mixins that inherit from it (TYPE_CHECKING)
 ):
     "Single-pass Tcl analyser assembled from mixin groups."
